# Should not substitute binds when `prepared_statements: true`