@matthewd 2020-05-10
Huh, it seems surprising to me that the prepared version is that much slower. 

Either way, a consistent +50% seems fair, and I do [still] think the binds are important in prepared-statements mode.

How does this performance compare to current master? I assume master is closer to the no-prepared-statements version for both values, and so this will be a slow-down for prepared... but does it affect prepared=false?