@kamipo 2020-05-10
Here is current master.

`prepared_statements: true`:

```
Warming up --------------------------------------
      where with ids    11.000  i/100ms
Calculating -------------------------------------
      where with ids    109.497  (± 8.2%) i/s -    550.000  in   5.064226s
```

`prepared_statements: false`:

```
Warming up --------------------------------------
      where with ids    11.000  i/100ms
Calculating -------------------------------------
      where with ids    109.824  (± 8.2%) i/s -    550.000  in   5.048983s
```

> How does this performance compare to current master? I assume master is closer to the no-prepared-statements version for both values, and so this will be a slow-down for prepared... but does it affect prepared=false?

I've tried the benchmark a few times locally, it will be a slow down for prepared about 15% ~ 20% (it is still about 25% faster than before though), looks like there is a bit overhead for prepared=false, but the performance is mostly same-ish.