@kamipo 2020-05-18
I'm going to merge this to fix substitute binds in `prepared_statements: true` as a regression.